# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an official Nova CDN API documentation.",
        "title": "Nova CDN API",
        "termsOfService": "https://novaardiansyah.my.id/live/nova-app/terms-of-service",
        "contact": {
            "name": "API Support",
            "url": "https://novaardiansyah.my.id",
            "email": "support@novadev.my.id"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/novaardiansyah/nova-cdn/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/galleries": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated list of galleries",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "galleries"
                ],
                "summary": "List galleries",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Subject ID",
                        "name": "subject_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Subject Type",
                        "name": "subject_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Size (original, small, medium, large)",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/nova-cdn_pkg_utils.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_controllers.GallerySwagger"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/nova-cdn_pkg_utils.ErrorSimpleResponse"
                        }
                    }
                }
            }
        },
        "/galleries/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload a new image to the gallery with optimized versions",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "galleries"
                ],
                "summary": "Upload image to gallery",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image file to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Subject ID",
                        "name": "subject_id",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Subject Type",
                        "name": "subject_type",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "default": "gallery",
                        "description": "Directory name (gallery, payment, item)",
                        "name": "dir",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Image description",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Set image as private",
                        "name": "is_private",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/nova-cdn_pkg_utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_controllers.GallerySwagger"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/nova-cdn_pkg_utils.ErrorSimpleResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/nova-cdn_pkg_utils.ErrorSimpleResponse"
                        }
                    }
                }
            }
        },
        "/galleries/{group_code}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Show galleries by group code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "galleries"
                ],
                "summary": "Show galleries by group code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group Code",
                        "name": "group_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Size (original, small, medium, large)",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/nova-cdn_pkg_utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_controllers.GallerySwagger"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/nova-cdn_pkg_utils.ErrorSimpleResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/nova-cdn_pkg_utils.ErrorSimpleResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Soft delete all galleries with the specified group code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "galleries"
                ],
                "summary": "Soft delete galleries by group code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group Code",
                        "name": "group_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Size (original, small, medium, large)",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/nova-cdn_pkg_utils.SimpleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/nova-cdn_pkg_utils.ErrorSimpleResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/nova-cdn_pkg_utils.ErrorSimpleResponse"
                        }
                    }
                }
            }
        },
        "/galleries/{group_code}/force": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently delete all galleries with the specified group code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "galleries"
                ],
                "summary": "Permanently delete galleries by group code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group Code",
                        "name": "group_code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Size (original, small, medium, large)",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/nova-cdn_pkg_utils.SimpleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/nova-cdn_pkg_utils.ErrorSimpleResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/nova-cdn_pkg_utils.ErrorSimpleResponse"
                        }
                    }
                }
            }
        },
        "/galleries/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Show a gallery item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "galleries"
                ],
                "summary": "Show a gallery item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Gallery ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/nova-cdn_pkg_utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_controllers.GallerySwagger"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/nova-cdn_pkg_utils.ErrorSimpleResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/nova-cdn_pkg_utils.ErrorSimpleResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Move a gallery item to trash",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "galleries"
                ],
                "summary": "Delete a gallery item (Soft Delete)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Gallery ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/nova-cdn_pkg_utils.SimpleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/nova-cdn_pkg_utils.ErrorSimpleResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/nova-cdn_pkg_utils.ErrorSimpleResponse"
                        }
                    }
                }
            }
        },
        "/galleries/{id}/force": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently delete a gallery item and its physical files",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "galleries"
                ],
                "summary": "Permanently delete a gallery item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Gallery ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/nova-cdn_pkg_utils.SimpleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/nova-cdn_pkg_utils.ErrorSimpleResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/nova-cdn_pkg_utils.ErrorSimpleResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_controllers.GallerySwagger": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_private": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "nova-cdn_pkg_utils.ErrorSimpleResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "nova-cdn_pkg_utils.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/nova-cdn_pkg_utils.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "nova-cdn_pkg_utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter the token with the `Bearer ` prefix, e.g. \"Bearer abcde12345\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}